requirements.txt
torch
torchvision
opencv-python
timm
Pillow
numpy
ffmpeg-python
